StoreHelper has been centralized into BaseStore [cis-622]
