+++ ZombieGame/Assets/Scripts/Stores/BaseStore.cs
@@ -1,5 +1,8 @@
+using System;
 using Assets.Scripts.Constants.Names;
 using Assets.Scripts.Constants.Types;
+using Assets.Scripts.Player;
+using Assets.Scripts.Weapons;
 using UnityEngine;
 
 namespace Assets.Scripts.Stores
@@ -11,20 +14,56 @@ namespace Assets.Scripts.Stores
         public abstract int AmmoReplenish { get; }
         public abstract WeaponType Type { get; }
 
-        public StoreHelper Store { get; private set; }
-
         private TextMesh TextMesh;
-        private AudioSource AmmoPurchaseSound;
+        private AudioSource PurchaseSound;
+
+        public int BuyAmmo(ref int Points)
+        {
+            if (Points < CostForAmmo)
+            {
+                return 0;
+            }
+            else
+            {
+                Points -= CostForAmmo;
+                PurchaseSound.Play();
+
+                return AmmoReplenish;
+            }
+        }
+
+        public Weapon PurchaseWeapon(ref int Points)
+        {
+            if (Points < CostToBuy)
+            {
+                return null;
+            }
+            else
+            {
+                Points -= CostToBuy;
+                PurchaseSound.Play();
+
+                switch (Type)
+                {
+                    case WeaponType.Pistol:
+                        return GameObject.Find(Enum.GetName(typeof(WeaponType), Type)).GetComponent<PlayerPistol>().Weapon; // TODO: These player objects will need a base class, then this can be shrunk 
+                    case WeaponType.Rifle:
+                        return GameObject.Find(Enum.GetName(typeof(WeaponType), Type)).GetComponent<PlayerRifle>().Weapon;
+                    case WeaponType.Laser:
+                        return GameObject.Find(Enum.GetName(typeof(WeaponType), Type)).GetComponent<PlayerLaser>().Weapon;
+                    default:
+                        throw new NotImplementedException();
+                }
+            }
+        }
 
         private void Start()
         {
-            AmmoPurchaseSound = gameObject.GetComponent<AudioSource>();
-
-            Store = new StoreHelper(CostToBuy, CostForAmmo, AmmoReplenish, AmmoPurchaseSound, Type);
+            PurchaseSound = gameObject.GetComponent<AudioSource>();
 
             TextMesh = gameObject.GetComponentInChildren<TextMesh>();
             TextMesh.color = Color.clear;
-            TextMesh.text = $"Buy: {Store.BuyCost}\nAmmo: {Store.AmmoCost}\nPress 'B' to buy";
+            TextMesh.text = $"Buy: {CostToBuy}\nAmmo: {CostForAmmo}\nPress 'B' to buy";
         }
 
         private void OnTriggerStay2D(Collider2D collision)
