--- ZombieGame/Assets/Scripts/Stores/StoreHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7dce8444b4ba2094d9764d2a9aa47776
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
