+++ ZombieGame/Assets/Scripts/Weapons/LaserWeapon.cs
@@ -43,7 +43,7 @@ namespace Assets.Scripts.Weapons
             return LastBulletShot == null;
         }
 
-        public override void PurchaseAmmo(ref int Points, StoreHelper store)
+        public override void PurchaseAmmo(ref int Points, BaseStore store)
         {
             if (store.Type == Type)
             {
